+++ demo/components/MaskedInput/BasicUsage.tsx
@@ -12,7 +12,6 @@ export const BasicUsage = (storyProps: StoryProps): React.ReactElement => {
         name="test"
         hasClearButton
         mask="+7 (###)-###-##-##"
-        // defaultValue={'9992024059'}
         placeholder="+7 (___)-___-__-__"
         onEnterPress={({ component }) => {
           console.log(component.name, component.name);