+++ demo/cypress/MaskedInput/index.tsx
@@ -4,13 +4,12 @@ import * as L from '../../../korus-ui';
 export const MaskedInput = (): React.ReactElement => {
   const [cardValue, setCardValue] = React.useState<string>('');
   const [phoneValue, setPhoneValue] = React.useState<string | null>('8002000600');
+  const [cardNumberValue, setCardNumberValue] = React.useState<string | null>('');
   const [isDisabled, setIsDisabled] = React.useState<boolean>(false);
-
-  const [inputValue, setInputValue] = React.useState<string | null>('+7 (800)-200-06-00');
 
   return (
     <L.Div _demoStory>
-      <L.Span>Номер телефона (не контролируемый с валидацией)</L.Span>
+      <L.Span>Phone number (uncontrolled with validation)</L.Span>
       <L.MaskedInput
         mask="+7 (###)-###-##-##"
         hasClearButton
@@ -21,7 +20,12 @@ export const MaskedInput = (): React.ReactElement => {
         name="MINotControlledPhone"
         form="my-form"
       />
-      <L.Span>СНИЛС (контролируемый)</L.Span>
+      <L.Switcher value={isDisabled} onChange={(event) => setIsDisabled(event.component.value)}>
+        Toggle isDisabled
+      </L.Switcher>
+      <br />
+      <br />
+      <L.Span>Insurance number (controlled)</L.Span>
       <L.MaskedInput
         mask="###-###-### ##"
         name="MIControlledInsurance"
@@ -29,7 +33,7 @@ export const MaskedInput = (): React.ReactElement => {
         value={cardValue}
         onChange={(event) => setCardValue(event.component.value)}
       />
-      <L.Span>Номер телефона (контролируемый)</L.Span>
+      <L.Span>Phone number (controlled)</L.Span>
       <L.MaskedInput
         name="MIControlledPhone"
         mask="+7 (###)-###-##-##"
@@ -37,39 +41,43 @@ export const MaskedInput = (): React.ReactElement => {
         value={phoneValue}
         onChange={(event) => {
           setPhoneValue(event.component.value);
-          console.log(event.component);
-          setInputValue(event.component.inputValue);
+          console.log(event.component.inputValue);
         }}
         onEnterPress={(event) => {
-          setInputValue(event.component.inputValue);
-          console.log(event.component);
+          console.log(event.component.inputValue);
         }}
       />
-      <L.Span>Номер машины (не контролируемый)</L.Span>
+      <L.Button _warning name="clearPhoneValue" onClick={() => {setPhoneValue(null)}}>Clear Value</L.Button>
+      {' '}
+      <L.Button _warning name="setPhoneValue" onClick={() => {setPhoneValue('9818862798')}}>Set Value</L.Button>
+      <br />
+      <br />
+      <L.Span>Credit card number (controlled without placeholder)</L.Span>
+      <L.MaskedInput
+        mask="####-####-####-####"
+        name="MICardNumberControlled"
+        value={cardNumberValue}
+        onChange={(event) => {
+          setCardNumberValue(event.component.value);
+          console.log(event.component.inputValue);
+        }}
+      />
+      <L.Button _warning name="clearCardValue" onClick={() => {setCardNumberValue(null)}}>Clear Value</L.Button>
+      <br />
+      <br />
+      <L.Span>Car Number (uncontrolled)</L.Span>
       <L.MaskedInput
         mask="LL##LL####"
         name="MICarNumber"
         placeholder="Car number"
       />
-      <L.Span>Номер кредитки (не контролируемый с начальным значением)</L.Span>
+      <L.Span>Credit card number (uncontrolled with default value)</L.Span>
       <L.MaskedInput
         mask="####-####-####-####"
         name="MICreditCardNumber"
         placeholder="####-####-####-####"
         defaultValue="6666777788889999"
       />
-      <br />
-      <br />
-      <L.Div name="InputValue">InputValue: {inputValue}</L.Div>
-      <br />
-      <br />
-      <L.Button name="clearValue" onClick={() => {setPhoneValue(null)}}>Clear Value</L.Button>
-      {' '}
-      <L.Button name="setValue" onClick={() => {setPhoneValue('9818862798')}}>Set Value</L.Button>
-      {' '}
-      <L.Switcher value={isDisabled} onChange={(event) => setIsDisabled(event.component.value)}>
-        Toggle isDisabled
-      </L.Switcher>
     </L.Div>
   );
 };