+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 - MaskedInput: Fixed backspace
 - Fixed test coverage report for Windows
+- MaskedInput: Remove mask on reset